refactor: move Playslip
