+++ src/test/java/step2/PlayslipTest.java
@@ -6,14 +6,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static step2.LotteryNumberTest.LotteryNumber;
-import static step2.NumberSelectionTest.NaturalSelection;
 
 public class PlayslipTest {
 
@@ -56,32 +52,4 @@ public class PlayslipTest {
     }
 
 
-    public static class Playslip {
-        public static final String ONLY_POSITIVE_NUMBERS = "선택은 1개 이상만 가능합니다.";
-        public static final List<Integer> NUMBER_POOL = makeNumberPool();
-        private static final int MAX_SELECT_NUMBER = 45;
-        public static final int SELECTION_COUNT = 6;
-
-        private final NaturalSelection selection = new NaturalSelection();
-
-        public List<LotteryNumber> selectNumbers(int numberSet) {
-            if (numberSet < 1) {
-                throw new IllegalArgumentException(ONLY_POSITIVE_NUMBERS);
-            }
-
-            List<LotteryNumber> result = new ArrayList<>();
-            for (int i = 0; i < numberSet; i++) {
-                result.add(new LotteryNumber(selection.select(NUMBER_POOL, SELECTION_COUNT)));
-            }
-            return result;
-        }
-
-        private static List<Integer> makeNumberPool() {
-            List<Integer> selectableNumberSource = new ArrayList<>();
-            for (int i = 0; i < MAX_SELECT_NUMBER; i++) {
-                selectableNumberSource.add(i + 1);
-            }
-            return Collections.unmodifiableList(selectableNumberSource);
-        }
-    }
 }
